Fix all 69 Maven compilation errors from plugin merge
- WeaponManager: cast raw Map.getOrDefault() to int
- GTMPlugin: pass `this` to GTM economy commands; register WatchlistMessageListener via plugin messenger
- InvSeeListener/PerksManager: add `implements Listener`
- MaintenanceManager: fix raw List<> for bypass UUIDs
- VanishCommand/BackPackNew/CommandSpy/InvSee: fix SimpleFormatter(this){} → SimpleFormatter(){}
- VanishCommand/BackPackNew/CommandSpy/InvSee/GangPlugin: fix ArrayList<CallSite> → ArrayList<String>
- VanishManager/FixCommand/FixAllCommand/FeedCommand: add missing SpigotAdminSystem import
- JetPackSystem: fix BukkitRunnable(this){} → BukkitRunnable(){} with outer-class refs; fix PersistentDataContainer.set byte cast
- LevelSystem/ScoreboardHandler: fix BukkitRunnable(this){} → BukkitRunnable(){} with outer-class refs
- TearGas/SmokeGrenade/ThrowableGrenade/WeaponListener: fix BukkitRunnable(this){} → BukkitRunnable(){}
- GangPlugin: fix Map.Entry<UUID,Object> to correct typed generics
- BackPackNew: fix Map.Entry.comparingByValue() type witness; fix iterator reuse as Object
- UtilsConfig/TimeHandler: fix setGameRule((Object)false) → setGameRule(false)
- Economy: fix ServicesManager.register (Object) cast
- SpigotAdminSystem: pass startTime instead of this to GCCommand
- WatchlistSpigot: fix WatchlistManager/WatchlistCommand/WatchlistMessageListener constructor args
- TradePlugin: fix TradeCommand/TradeListener constructor args
- ClearChatCommand: fix Object player → Player player

+++ GTM-Plugin/src/main/java/Guns/WeaponManager.java
@@ -77,19 +77,19 @@ public class WeaponManager {
     }
 
     public static int getAmmo(Player player, AmmoType ammoType) {
-        return ((Map)playerAmmo.getOrDefault(player.getUniqueId(), new EnumMap(AmmoType.class))).getOrDefault((Object)ammoType, 0);
+        return (int)((Map)playerAmmo.getOrDefault(player.getUniqueId(), new EnumMap(AmmoType.class))).getOrDefault((Object)ammoType, 0);
     }
 
     public static void decreaseAmmo(Player player, AmmoType ammoType, int amount) {
         Map ammoMap = playerAmmo.computeIfAbsent(player.getUniqueId(), k -> new EnumMap(AmmoType.class));
-        int newAmount = ammoMap.getOrDefault((Object)ammoType, 0) - amount;
+        int newAmount = (int)ammoMap.getOrDefault((Object)ammoType, 0) - amount;
         ammoMap.put(ammoType, newAmount);
         WeaponManager.saveAmmoData();
     }
 
     public static void addAmmo(Player player, AmmoType ammoType, int amount) {
         Map ammoMap = playerAmmo.computeIfAbsent(player.getUniqueId(), k -> new EnumMap(AmmoType.class));
-        int newAmount = ammoMap.getOrDefault((Object)ammoType, 0) + amount;
+        int newAmount = (int)ammoMap.getOrDefault((Object)ammoType, 0) + amount;
         ammoMap.put(ammoType, newAmount);
         WeaponManager.saveAmmoData();
     }
